make changes to tests in the learning-object-route-handler so they all  pass in docker, improve the structe of mocks and fix any failing tests throughout the service
