+++ package.json
@@ -57,20 +57,10 @@
     "@types/request-promise": "^4.1.39",
     "@types/striptags": "^3.1.1",
     "@types/supertest": "^2.0.7",
-<<<<<<< HEAD
-    "@types/tape": "^4.2.31",
-    "@types/uuid": "^3.4.4",
-    "gulp": "^3.9.1",
-    "gulp-nodemon": "^2.2.1",
-    "gulp-typescript": "^4.0.2",
-    "jest": "^23.4.0",
-    "jsonwebtoken": "^8.5.0",
-=======
     "gulp": "^4.0.0",
     "gulp-nodemon": "^2.4.2",
     "gulp-typescript": "^5.0.1",
     "jest": "^24.5.0",
->>>>>>> master
     "mongodb-memory-server": "^2.7.4",
     "supertest": "^3.3.0",
     "ts-jest": "^24.0.0",
