+++ src/LearningObjects/LearningObjectRouteHandler.spec.ts
@@ -1,5 +1,6 @@
 import { MongoDriver } from '../drivers/MongoDriver';
-import { MOCK_OBJECTS, generateToken } from '../tests/mocks';
+import { MOCK_OBJECTS} from '../tests/mocks';
+import { generateToken } from '../tests/mock-token-manager';
 import * as LearningObjectRouteHandler from './LearningObjectRouteHandler';
 import * as express from 'express';
 import * as bodyParser from 'body-parser';
@@ -8,16 +9,14 @@ import { MockLibraryDriver } from '../tests/mock-drivers/MockLibraryDriver';
 import { MockS3Driver } from '../tests/mock-drivers/MockS3Driver';
 import { LibraryCommunicator, FileManager } from '../interfaces/interfaces';
 import * as cookieParser from 'cookie-parser';
-import { processToken } from '../middleware';
+import { processToken, handleProcessTokenError } from '../middleware';
 
 const app = express();
 const router = express.Router();
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(cookieParser());
-app.use(processToken, (error: any, req: any, res: any, next: any) =>
-    next(),
-);
+app.use(processToken, handleProcessTokenError);
 app.use(router);
 const request = supertest(app);
 const testObjectID = 'parent_object_1';
@@ -46,7 +45,6 @@ describe('LearningObjectRouteHandler', () => {
         it('should return a learing object based on the id', done => {
             request
                 .get(`/learning-objects/${testObjectID}`)
-                .expect('Content-Type', /json/)
                 .expect(200)
                 .then(res => {
                     expect(res.text).toContain(`${testObjectID}`);
@@ -56,7 +54,7 @@ describe('LearningObjectRouteHandler', () => {
         it('should return a status of 500 and an Error message', done => {
             request
                 .get(`/learning-objects/${testObjectID}123`)
-                .expect('Content-Type', /json/)
+
                 .expect(500)
                 .then(res => {
                     expect(res).toBeDefined();
@@ -68,7 +66,7 @@ describe('LearningObjectRouteHandler', () => {
         it('should return the materials for the specified learning object', done => {
             request
                 .get(`/learning-objects/${testObjectID}/materials/all`)
-                .expect('Content-Type', /json/)
+
                 .expect(200)
                 .then(res => {
                     expect(res.text).toContain('url');
@@ -81,9 +79,8 @@ describe('LearningObjectRouteHandler', () => {
         it('should update the requested learning object and return a status of 200', done => {
             request
                 .patch(`/learning-objects/${testObjectID}`)
+                .set('Authorization', `Bearer ${token}`)
                 .send({ learningObject: { name: 'Java stuff' } })
-                .set(authorization)
-                .expect('Content-Type', /text/)
                 .expect(200)
                 .then(res => {
                     done();
@@ -96,7 +93,7 @@ describe('LearningObjectRouteHandler', () => {
             request
                 .get(`/learning-objects/${testObjectID}/children/summary`)
                 .expect(200)
-                .expect('Content-Type', /json/)
+
                 .then(res => {
                     expect(res.text).toContain('author');
                     done();
@@ -108,7 +105,7 @@ describe('LearningObjectRouteHandler', () => {
         it('should delete a learning object from the database and return a 200 status', done => {
             request
                 .delete(`/learning-objects/Java%20stuff`)
-                .set(authorization)
+                .set('Authorization', `Bearer ${token}`)
                 .expect(200)
                 .then(() => {
                     done();
