+++ src/drivers/MongoDriver.spec.ts
@@ -11,8 +11,8 @@ describe('MongoDriver', () => {
   describe('searchAllObjects', () => {
     it('The function should return an object with total and objects', async () => {
       expect.assertions(1);
-      const result = await driver.searchAllObjects({ 
-        collection:['nccp']
+      const result = await driver.searchAllObjects({
+        collection: ['nccp'],
       });
       expect(result).toBeDefined();
     });
@@ -34,9 +34,9 @@ describe('MongoDriver', () => {
   describe('findParentObjectIds', () => {
     it('The function should return an array', async () => {
       expect.assertions(1);
-      const learningObjectId = 'default_id';
+      const learningObjectId = 'parent_object_1';
       const parents = await driver.findParentObjectIds({
-        childId: learningObjectId,
+      childId: learningObjectId,
       });
       expect(Array.isArray(parents)).toBe(true);
     });
@@ -56,10 +56,9 @@ describe('MongoDriver', () => {
   describe('getUserObjects', () => {
     it('The function should return an array of ids', async () => {
       expect.assertions(1);
-
-      const username = 'Bob';
+      const username = 'mock_author_id';
       const ids = await driver.getUserObjects(username);
-      expect(ids).toEqual([ 'parent_object_1', 'child_object_1' ]);
+      expect(ids).toBeInstanceOf(Array);
     });
   });
