+++ src/drivers/SentryConnector.ts
@@ -2,4 +2,5 @@ import * as raven from 'raven';
 
 export const reportError = (error: Error) => {
   raven.captureException(error);
+  console.error(error);
 };
